apollo_l1_provider: move l1 provider config to dedicated config file
