+++ crates/apollo_integration_tests/src/utils.rs
@@ -28,8 +28,8 @@ use apollo_l1_endpoint_monitor::monitor::L1EndpointMonitorConfig;
 use apollo_l1_gas_price::eth_to_strk_oracle::{EthToStrkOracleConfig, ETH_TO_STRK_QUANTIZATION};
 use apollo_l1_gas_price::l1_gas_price_provider::L1GasPriceProviderConfig;
 use apollo_l1_gas_price_types::DEFAULT_ETH_TO_FRI_RATE;
+use apollo_l1_provider::config::L1ProviderConfig;
 use apollo_l1_provider::l1_scraper::L1ScraperConfig;
-use apollo_l1_provider::L1ProviderConfig;
 use apollo_mempool::config::MempoolConfig;
 use apollo_mempool_p2p::config::MempoolP2pConfig;
 use apollo_monitoring_endpoint::config::MonitoringEndpointConfig;
