+++ crates/apollo_l1_provider/tests/reexecution_flow.rs
@@ -1,8 +1,9 @@
 use std::sync::Arc;
 
 use apollo_batcher_types::communication::MockBatcherClient;
+use apollo_l1_provider::config::L1ProviderConfig;
 use apollo_l1_provider::l1_provider::L1ProviderBuilder;
-use apollo_l1_provider::{L1ProviderConfig, ProviderState};
+use apollo_l1_provider::ProviderState;
 use apollo_l1_provider_types::MockL1ProviderClient;
 use apollo_state_sync_types::communication::MockStateSyncClient;
 use apollo_time::test_utils::FakeClock;
